+++ case1/popular-movies/package.json
@@ -9,6 +9,7 @@
     "axios": "^0.23.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-router-dom": "^5.3.0",
     "react-scripts": "4.0.3",
     "styled-components": "^5.3.3",
     "web-vitals": "^1.0.1"
