+++ case1/popular-movies/src/index.js
@@ -3,10 +3,13 @@ import ReactDOM from 'react-dom';
 import './style.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
+import { Route, Router, Switch } from 'react-router-dom';
 
 ReactDOM.render(
   <React.StrictMode>
-    <App />
+  
+      <App />
+
   </React.StrictMode>,
   document.getElementById('root')
 );
